[filenames] Create a separate filename linter
Summary:
We were using the ArcanistFilenameLinter, but it rejected plus signs
in filenames. This is legit in objective-c filenames, so we should just use
our own filename linter. Using this approach instead of extending the ArcanistFilenameLinter
because it's a final class.

Test Plan: arc lint something+somethingelse.h and see that there are no lint errors.

Reviewers: bobo, vlad

Reviewed By: bobo

Maniphest Tasks: T701

Differential Revision: https://phabricator.chronosresearch.com/D80