[ios] Disable Bad Charset Linter
Summary:
Xcode is totally fine with unicode in the .h/.m files, as evidenced
by the new copyright symbols it automatically adds to the top of new files. We
don't need this linter anymore.

Also fixing the linter for cassowary itself, since we merged this change incorrectly
from the upstream.

Test Plan:
Add a comment with a copyright symbol, run `arc lint`. No more
errors or warnings.

Reviewers: chris

Reviewed By: chris

Differential Revision: https://phabricator.robinhood.com/D6315